fix file extensions
