+++ src/components/Body.jsx
@@ -9,4 +9,3 @@ export default function Body(props){
         </div>
     );
 }
-
