+++ src/components/MobileMenu/index.module.css
@@ -1,0 +1,54 @@
+.container {
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100vh;
+    width: 100vw;
+    z-index: 99;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-items: center;
+
+    font-size: 4em;
+    text-align: center;
+    color: white;
+    background-color: rgba(0,0,0,0.8);
+}
+
+.hidden {
+    z-index: -99;
+    display: none;
+}
+
+
+.item {
+    margin: 10px;
+    text-decoration: none;
+    z-index: 100;
+}
+
+.item:first-child{
+    margin-top: 2em;
+}
+
+.item:link{
+    color: white;
+}
+
+.item:visited{
+    color: white;
+}
+
+.item:active{
+    color: #e03997;
+}
+
+
+.icon {
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    margin: 0;
+}
