+++ src/components/splash/SplashImage.jsx
