+++ src/layouts/SplashLayout.jsx
