+++ src/pages/404.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Template from '../layouts/Template';
-import Menu from '../components/Menu.js';
-import Body from '../components/Body.js';
+import Menu from '../components/Menu';
+import Body from '../components/Body';
 
 export default function Missing404(props) {
