+++ src/pages/gallery.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import Template from '../layouts/Template';
-import Menu from '../components/Menu.js';
-import Body from '../components/Body.js';
+import Menu from '../components/Menu';
+import Body from '../components/Body';
 
-export default function Contact(props) {
+export default function Gallery(props) {
     return (
         <div>
             <Template>
                 <Menu />
                 <Body className="white center-text">
-                    <h1 className="title-font large-text">Contact</h1>
-                    <p>Email the big dog baritone</p>
+                    <h1 className="title-font large-text">Gallery</h1>
+                    <p>Look at the big dog baritone</p>
                 </Body>
             </Template>
         </div>
