+++ src/pages/listen.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { graphql } from 'gatsby';
 import get from 'lodash/get';
 import Template from '../layouts/Template';
-import Menu from '../components/Menu.js';
-import Body from '../components/Body.js';
+import Menu from '../components/Menu';
+import Body from '../components/Body';
 
 export default function Listen(props) {
     const videos = get(props, 'data.allMarkdownRemark.edges');
@@ -15,19 +15,13 @@ export default function Listen(props) {
                 <Body className="white center-text">
                     <h1 className="title-font large-text">Listen</h1>
                     <div id="Videos">
-                        {videos.map(({node}) => {
+                        {videos.map(({ node }) => {
                             const description = get(
                                 node,
                                 'frontmatter.description'
                             );
-                            const slug = get(
-                                node,
-                                'fields.slug'
-                            );
-                            const html = get(
-                                node,
-                                'html'
-                            );
+                            const slug = get(node, 'fields.slug');
+                            const html = get(node, 'html');
 
                             return (
                                 <div key={slug}>
@@ -48,22 +42,22 @@ export default function Listen(props) {
 }
 
 export const videosQuery = graphql`
-           query VideosQuery {
-               allMarkdownRemark(
-                   filter: { fileAbsolutePath: { regex: "/videos/" } }
-                   sort: { order: ASC, fields: fields___slug }
-               ) {
-                   edges {
-                       node {
-                           fields {
-                               slug
-                           }
-                           html
-                           frontmatter {
-                               description
-                           }
-                       }
-                   }
-               }
-           }
-       `;
+    query VideosQuery {
+        allMarkdownRemark(
+            filter: { fileAbsolutePath: { regex: "/videos/" } }
+            sort: { order: ASC, fields: fields___slug }
+        ) {
+            edges {
+                node {
+                    fields {
+                        slug
+                    }
+                    html
+                    frontmatter {
+                        description
+                    }
+                }
+            }
+        }
+    }
+`;
